% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clavaan.r
\name{ccfa}
\alias{ccfa}
\title{Run CFA model with censored data}
\usage{
ccfa(model = NULL, data = NULL, bounds = NULL, ...)
}
\arguments{
\item{model}{lavaan model}

\item{data}{data frame}

\item{bounds}{censored points}

\item{...}{Additional lavaan arguments}
}
\value{
`ccfa()` returns `lavaan` class.
}
\description{
`ccfa()`
}
