% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_3_dcMultivariate.r
\name{cMulti}
\alias{cMulti}
\title{Compute mean and covariance matrix of censored data}
\usage{
cMulti(data, bounds, scaling = F)
}
\description{
Compute mean and covariance matrix of censored data
}
