% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clavaan.r
\name{csem}
\alias{csem}
\title{Run SEM model with censored data}
\usage{
csem(model = NULL, data = NULL, bounds = NULL, ...)
}
\arguments{
\item{model}{lavaan model}

\item{data}{data frame}

\item{bounds}{censored points}

\item{...}{Additional lavaan arguments}
}
\value{
`csem()` returns `lavaan` class.
}
\description{
`csem()`
}
