% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_2_dcmMLE.r
\name{ll_censored_bi}
\alias{ll_censored_bi}
\title{Compute likelihood of Bivariate censored data}
\usage{
ll_censored_bi(theta, XY, bounds, fixed = NULL)
}
\description{
Compute likelihood of Bivariate censored data
}
